style(graphin):adjust demos
