+++ packages/graphin/docs/render/element/edge.md
@@ -1,0 +1,13 @@
+---
+title: 内置边
+group:
+  path: /render
+nav:
+  title: Graphin
+  path: /graphin
+  order: 1
+---
+
+## 基本介绍
+
+<code src='./demos/edge.tsx'>
