+++ packages/graphin/docs/render/element/node.md
@@ -1,0 +1,13 @@
+---
+title: 节点元素
+group:
+  path: /render
+nav:
+  title: Graphin
+  path: /graphin
+  order: 1
+---
+
+## 基本介绍
+
+<code src='./demos/node.tsx'>
